refine the room cleanup
